Rename: ProfieImg컴포넌트 common하위로 이동
