+++ webapp/src/components/UserCard/index.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { userCardType } from 'types/user.type';
-import ProfileImg from 'components/ProfileImg';
+import ProfileImg from 'components/Common/ProfileImg';
 import { userCardParser } from 'service/user/user.parser';
 import TechSkills from 'components/TechSkills';
 import HeartSvg from 'assets/icons/HeartSvg';
